Fix crash in validating MultiCurrencyAmounts
There was a force unwrap which crashed in certain conditions.
Basically it assumed that all currencies in a transaction were
used as amount as well - this is true for balanced transactions,
but e.g. if the transaction has only one posting with a price
in another currency the method would crash

Add test to catch this in the future