Revert "Change transactions, plugins, events and customs to Set"
Transactions can be duplicated, so cannot be a Set. For the rest
it is nice to keep the order to e.g. not mix it up when writing it out.

Make a some things comparable to allow comparing the contents of these
arrays without taking the order into account

Own class for Option

This reverts commit 916cea5dae88ee4aaf7a0c799cd060104d834739.